# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API TFilms",
        "version": "1.0"
    },
    "paths": {
...
        "/film": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Получение списка фильмов",
                "description": "Получение списка фильмов",
                "operationId": "aeb112c4a8c96e5905cbad92a69405b4",
                "requestBody": {
                    "$ref": "#/components/requestBodies/FilmGetRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FilmResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": "syntax error, unexpected token \"}\""
                                                },
                                                "Info": {
                                                    "$ref": "#/components/schemas/Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "film"
                ],
                "summary": "Создание фильма",
                "description": "Создание фильма",
                "operationId": "9ed355b538d75c8acd3b880c5d99540a",
                "requestBody": {
                    "$ref": "#/components/requestBodies/FilmSaveRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": "syntax error, unexpected token \"}\""
                                                },
                                                "Info": {
                                                    "$ref": "#/components/schemas/Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "film"
                ],
                "summary": "Удаление фильма. На самом деле это постзапрос с параметром в теле запроса method=delete",
                "description": "Удаление фильма. На самом деле это постзапрос с параметром в теле запроса method=delete",
                "operationId": "027ec193f0a77c6f94d91d800d9bfa40",
                "requestBody": {
                    "$ref": "#/components/requestBodies/FilmDeleteRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": "syntax error, unexpected token \"}\""
                                                },
                                                "Info": {
                                                    "$ref": "#/components/schemas/Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "patch": {
                "tags": [
                    "film"
                ],
                "summary": "Изменение фильма. На самом деле это постзапрос с параметром в теле запроса method=patch",
                "description": "Изменение фильма. На самом деле это постзапрос с параметром в теле запроса method=patch",
                "operationId": "2483e9d02f7af3312de5a12dec5ecefc",
                "requestBody": {
                    "$ref": "#/components/requestBodies/FilmChangeRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "format": "int",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": "syntax error, unexpected token \"}\""
                                                },
                                                "Info": {
                                                    "$ref": "#/components/schemas/Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/film/{film_id}": {
            "get": {
                "tags": [
                    "film"
                ],
                "summary": "Получение одного фильма",
                "description": "Получение одного фильма",
                "operationId": "7b31b18db3ab6a34e1b336aebc73f5d0",
                "requestBody": {
                    "$ref": "#/components/requestBodies/FilmGetRequest"
                },
                "responses": {
                    "200": {
                        "description": "Успешно",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/FilmResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "format": "string",
                                            "example": "Error"
                                        },
                                        "data": {
                                            "properties": {
                                                "Message": {
                                                    "type": "string",
                                                    "format": "string",
                                                    "example": "syntax error, unexpected token \"}\""
                                                },
                                                "Info": {
                                                    "$ref": "#/components/schemas/Error"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error": {
                "properties": {
                    "line": {
                        "type": "integer",
                        "example": 37
                    },
                    "file": {
                        "type": "string",
                        "format": "string",
                        "example": "/path_to_file/..."
                    }
                },
                "type": "object"
            },
...
            "FilmResource": {
                "properties": {
                    "film_id": {
                        "type": "integer",
                        "format": "number",
                        "example": "1"
                    },
                    "name": {
                        "type": "string",
                        "format": "string",
                        "example": "Форсаж"
                    },
                    "photo": {
                        "type": "string",
                        "format": "string",
                        "example": "dgdsfg.png"
                    },
                    "description": {
                        "type": "string",
                        "format": "string",
                        "example": "Описание фильма форсаж"
                    },
                    "duration": {
                        "type": "string",
                        "format": "time",
                        "example": "02:12"
                    },
                    "age_restrictions": {
                        "type": "integer",
                        "format": "number",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
        },
        "requestBodies": {
...
            "FilmChangeRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "id": {
                                    "schema": "required|int",
                                    "type": "integer",
                                    "format": "number",
                                    "example": "1"
                                },
                                "name": {
                                    "schema": "string|min:3|max:255|unique:App\\Models\\Films,name",
                                    "type": "string",
                                    "format": "string",
                                    "example": "Форсаж"
                                },
                                "photo": {
                                    "schema": "file|mimes:jpeg,png|max:4096",
                                    "type": "file",
                                    "format": "image",
                                    "example": "sfaf.png"
                                },
                                "description": {
                                    "schema": "string|min:5|max:255",
                                    "type": "string",
                                    "format": "string",
                                    "example": "Описание фильма форсаж"
                                },
                                "duration": {
                                    "schema": "string",
                                    "type": "string",
                                    "format": "time",
                                    "example": "02:10"
                                },
                                "age_restrictions": {
                                    "schema": "int|string",
                                    "type": "integer",
                                    "format": "number",
                                    "example": "2"
                                },
                                "method": {
                                    "schema": "string",
                                    "type": "string",
                                    "format": "string",
                                    "example": "patch"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "FilmDeleteRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "id": {
                                    "schema": "required|int",
                                    "type": "integer",
                                    "format": "number",
                                    "example": "1"
                                },
                                "method": {
                                    "schema": "required|int",
                                    "type": "string",
                                    "format": "string",
                                    "example": "delete"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "FilmGetRequest": {
                "required": false,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "order": {
                                    "schema": "string:ASC,DESC",
                                    "type": "string",
                                    "format": "string",
                                    "example": "ASC"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "FilmSaveRequest": {
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "name": {
                                    "schema": "required|string|min:3|max:255|unique:App\\Models\\Films,name",
                                    "type": "string",
                                    "format": "string",
                                    "example": "Форсаж"
                                },
                                "photo": {
                                    "schema": "required|file|mimes:jpeg,png|max:4096",
                                    "type": "file",
                                    "format": "image",
                                    "example": "sfaf.png"
                                },
                                "description": {
                                    "schema": "required|string|min:5|max:255",
                                    "type": "string",
                                    "format": "string",
                                    "example": "Описание фильма форсаж"
                                },
                                "duration": {
                                    "schema": "required|string",
                                    "type": "string",
                                    "format": "time",
                                    "example": "02:10"
                                },
                                "age_restrictions": {
                                    "schema": "required|int|string",
                                    "type": "integer",
                                    "format": "number",
                                    "example": "2"
                                },
                                "method": {
                                    "schema": "required|string",
                                    "type": "string",
                                    "format": "string",
                                    "example": "post"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
...
        }
    },
...
}